refactor: rename GameRoomUpdateEvent to GameRoomUpdateMessage #75
